// lib/boxscore-types.ts
export interface BoxScoreResponse {
  game: BoxScoreGame;
}

export interface BoxScoreGame {
  gameId: string;
  gameCode: string;
  gameStatus: number;
  gameStatusText: string;
  period: number;
  gameClock: string;
  gameTimeUTC: string;
  gameEt: string;
  duration: number;
  attendance: number;
  sellout: string;
  arena: {
    arenaId: number;
    arenaName: string;
    arenaCity: string;
    arenaState: string;
    arenaCountry: string;
    arenaTimezone: string;
  };
  officials: {
    personId: number;
    name: string;
    nameI: string;
    firstName: string;
    familyName: string;
    jerseyNum: string;
    assignment: string;
  }[];
  homeTeam: BoxScoreTeam;
  awayTeam: BoxScoreTeam;
}

export interface BoxScoreTeam {
  teamId: number;
  teamName: string;
  teamCity: string;
  teamTricode: string;
  score: number;
  inBonus: string;
  timeoutsRemaining: number;
  periods: { period: number; periodType: string; score: number }[];
  players: BoxScorePlayer[];
  statistics: TeamStatistics;
}